Refactor AJAX button callback to return a button instead of text.
Previously, when clicking on the "Follow" button, the AJAX callback
would disable the button and would not allow the logged-in user to
unfollow the user.

Fixes #15.

This commit changes the AJAX callback to return a button instead of
text so the logged-in user can now "Unfollow" a user without reloading
the page.