Add support to mark notifications as unread.
This commit changes how notifications are cleared.

Previously, follow notifications were redirected to the logged-in
user's followers page and deleted immediately.  Now, follow
notifications are redirected to the follower's profile and
marked as read.

For the bulk follow notification ("X users are now
following you"), the logged-in user is now redirected to the
notifications unread page and not marked as read or deleted.  This
allows the logged-in user to view their follower notifications in the
date that they were followed.

Fixes #24.