Follow Blogs: Remove set_activity_score_on_activity_directory() method
This class method interfered with the 'bp_activity_screen_index' hook,
since this hook was used both as an action and a filter.  This is a bug
in WP core as it cannot distinguish between the two contexts when the hook
is named the same.

This functionality isn't being used anymore anyway in follow-blogs, so
we're removing it here.

Hat-tip @christianwach.

Fixes #47.