Follow Blogs: Support heartbeat in groupblog override query.
If you're on the activity directory and on the "Followed Sites" tab, the
"Load Newest" heartbeat block will show even when a new blog post has not
been added.

The issue is we need to look for the 'since' parameter in the activity
loop parameters and add this SQL statement to groupblog query.

See #48.